requirements.txt
# order is the order pip will install, so if package A depends on B make sure B comes first
setuptools>=17.1.1
# postgres python driver
psycopg2>=2.6.1
# ORM tool
SQLAlchemy>=1.0.11
# database migrations
alembic>=0.8.6
# support to yaml file format
PyYAML>=3.11
# json schema validator
jsonschema>=2.5.1
# support for ldap authentication
ldap3>=1.3.3
# rest-api
flask>=0.10.1
# manage db session from flask (requirement from potion)
Flask-SQLAlchemy>=2.1
# help create the endpoints for the rest-api
flask-potion>=0.12.6
